add final feat HW4
