+++ WBChatAppSUI/WBChatAppSUI.xcodeproj/xcuserdata/dariakorotkova.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,12 +7,17 @@
 		<key>WBChatAppSUI.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>1</integer>
+		</dict>
+		<key>WBChatWidgetExtension.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>2</integer>
 		</dict>
 		<key>WBWidgetExtension.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>1</integer>
+			<integer>0</integer>
 		</dict>
 	</dict>
 </dict>
