+++ WBChatAppSUI/WBChatAppSUI/Support/Info.plist
@@ -5,6 +5,7 @@
 	<key>NSUserActivityTypes</key>
 	<array>
 		<string>ConfigurationIntent</string>
+		<string>MyCustomIntentIntent</string>
 	</array>
 </dict>
 </plist>
