+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/AccentColor.colorset/Contents.json
