+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/AppIcon.appiconset/Contents.json
