+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Contents.json
