+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/DefPhoto.imageset/Change Avatar.png
