+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/DefPhoto.imageset/Contents.json
