+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Neutral_Active.colorset/Contents.json
