+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo1.imageset/Contents.json
