+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo2.imageset/1c3f407d0c7ff3bc9e52078ef2bbe372.jpeg
