+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo2.imageset/Contents.json
