+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo3.imageset/2c4ab2e6878b848cf34cec5937b90ffb.jpeg
