+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo3.imageset/Contents.json
