+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo4.imageset/Contents.json
