+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo6.imageset/Contents.json
