+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo7.imageset/Contents.json
