+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo7.imageset/a70ee65e1cab2c63136a619f1b00e1fe.jpeg
