+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo8.imageset/Contents.json
