+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/Photo9.imageset/Contents.json
