+++ WBChatAppSUI/WBChatWidget/Assets.xcassets/WidgetBackground.colorset/Contents.json
