+++ WBChatAppSUI/WBChatWidget/Info.plist
