+++ WBChatAppSUI/WBChatWidget/WBChatWidgetBundle.swift
@@ -1,0 +1,16 @@
+//
+//  WBChatWidgetBundle.swift
+//  WBChatWidget
+//
+//  Created by Daria on 21.06.2024.
+//
+
+import WidgetKit
+import SwiftUI
+
+@main
+struct WBChatWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        WBChatWidget()
+    }
+}
