--- WBChatAppSUI/WBWidget/WBWidgetBundle.swift
@@ -1,16 +1,0 @@
-//
-//  WBWidgetBundle.swift
-//  WBWidget
-//
-//  Created by Daria on 14.06.2024.
-//
-
-import WidgetKit
-import SwiftUI
-
-@main
-struct WBWidgetBundle: WidgetBundle {
-    var body: some Widget {
-        WBWidget()
-    }
-}
